# angularboot5/src/app/services/models/address-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AddressDto {
  city: string;
  country: string;
  houseNumber: number;
  id?: number;
  street: string;
  userId?: number;
  zipCode: number;
}