# angularboot5/src/app/services/models/transaction-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface TransactionDto {
  amount?: number;
  destinationIbn?: string;
  id?: number;
  transactionType?: 'TRANSFERT' | 'DEPOSIT';
  userId?: number;
}