# app_bancaire/src/main/java/com/lehcene/app_bancaire/dto/RoleResponse.java
package com.lehcene.app_bancaire.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class RoleResponse {
    private Integer id;
    private String name;
}